Modify regex for include statements
We may have multiple includes in following lines.
Currently the parser will only parse the follwoing pattern:
@include "onefile"
,
@include "anotherfile"

but throw an error with the following:
@include "onefile",
@include "anotherfile"

The latter pattern, although not explicitly mentioned in the format definition,
is supported by existing c++ and perl parsers. So it should be supported here as well.